Split serializers.py into resources
