+++ comicsdb/serializers/genre.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from comicsdb.models import Genre
+
+
+class GenreSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Genre
+        fields = ("id", "name")
