+++ comicsdb/serializers/rating.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from comicsdb.models import Rating
+
+
+class RatingSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Rating
+        fields = ("id", "name")
